Use `bencher` instead of unstable libtest
