+++ Cargo.toml
@@ -56,7 +56,6 @@ series = ["vtab"]
 # check for invalid query.
 extra_check = []
 modern_sqlite = ["libsqlite3-sys/bundled_bindings"]
-unstable = []
 in_gecko = ["modern_sqlite", "libsqlite3-sys/in_gecko"]
 bundled-windows = ["libsqlite3-sys/bundled-windows"]
 
@@ -82,6 +81,9 @@ lazy_static = "1.0"
 regex = "1.0"
 uuid = { version = "0.8", features = ["v4"] }
 unicase = "2.4.0"
+# Use `bencher` over criterion becasue it builds much faster and we don't have
+# many benchmarks
+bencher = "0.1"
 
 [dependencies.libsqlite3-sys]
 path = "libsqlite3-sys"
@@ -96,6 +98,10 @@ name = "deny_single_threaded_sqlite_config"
 
 [[test]]
 name = "vtab"
+
+[[bench]]
+name = "cache"
+harness = false
 
 [package.metadata.docs.rs]
 features = [ "backup", "blob", "chrono", "collation", "functions", "limits", "load_extension", "serde_json", "trace", "url", "vtab", "window", "modern_sqlite" ]
